+++ admin_site/system/migrations/0098_customer_site_prefix.py
@@ -4,7 +4,6 @@ from django.db import migrations, models
 
 
 class Migration(migrations.Migration):
-
     dependencies = [
         ("system", "0097_customer_using_sso"),
     ]