+++ admin_site/system/migrations/0107_alter_pc_name.py
@@ -6,7 +6,6 @@ import re
 
 
 class Migration(migrations.Migration):
-
     dependencies = [
         ("system", "0106_alter_pc_name_alter_site_uid"),
     ]