+++ admin_site/system/migrations/0108_alter_pc_name.py
@@ -6,7 +6,6 @@ import re
 
 
 class Migration(migrations.Migration):
-
     dependencies = [
         ("system", "0107_alter_pc_name"),
     ]